totempg.h: change type of totempg_group.group_len from int to size_t
* include/corosync/totem/totempg.h (struct totempg_group) [group_len]:
Change member type from int to size_t.

git-svn-id: http://svn.fedorahosted.org/svn/corosync/trunk@2036 fd59a12c-fef9-0310-b244-a6a79926bd2f

+++ include/corosync/totem/totempg.h
@@ -43,7 +43,7 @@
 
 struct totempg_group {
 	const void *group;
-	int group_len;
+	size_t group_len;
 };
 
 #define TOTEMPG_AGREED			0
